upgrades index division to python 3

+++ IMGdescriptors/Python/EdgeHistogramComputer.py
@@ -36,7 +36,7 @@ class EdgeHistogramComputer(DescriptorComputer):
 		for row in range(self.rows):
 			for col in range(self.cols):
 				mask = np.zeros_like(frame)
-				mask[((frameH/self.rows)*row):((frameH/self.rows)*(row+1)),(frameW/self.cols)*col:((frameW/self.cols)*(col+1))] = 255
+				mask[int((frameH/self.rows)*row):int((frameH/self.rows)*(row+1)), int(frameW/self.cols)*col:int((frameW/self.cols)*(col+1))] = 255
 				hist = cv2.calcHist([dominantGradients], [0], mask, self.bins, self.range)
 				hist = cv2.normalize(hist, None)
 				descriptor.append(hist)
