+++ IMGdescriptors/Python/OrientedGradientsComputer.py
@@ -28,7 +28,7 @@ class OrientedGradientsComputer(DescriptorComputer):
 		for row in range(self.rows):
 			for col in range(self.cols):
 				mask = np.zeros_like(frame)
-				mask[((frameH/self.rows)*row):((frameH/self.rows)*(row+1)),(frameW/self.cols)*col:((frameW/self.cols)*(col+1))] = 1
+				mask[int((frameH/self.rows)*row):int((frameH/self.rows)*(row+1)), int(frameW/self.cols)*col:int((frameW/self.cols)*(col+1))] = 1
 				mask[mask_threshold] = 0
 				a_, b_ = mask.shape
 				hist = cv2.calcHist([orientations], self.channel, mask, [self.bins], self.range)
